+++ awscli/paramfile.py
@@ -10,16 +10,38 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import copy
 import logging
 import os
 
 from awscli.compat import compat_open
+from awscli import argprocess
 
 logger = logging.getLogger(__name__)
 
 
+def register_uri_param_handler(session, **kwargs):
+    prefix_map = copy.deepcopy(LOCAL_PREFIX_MAP)
+    handler = URIArgumentHandler(prefix_map)
+    session.register('load-cli-arg', handler)
+
+
 class ResourceLoadingError(Exception):
     pass
+
+
+class URIArgumentHandler(object):
+    def __init__(self, prefixes):
+        self._prefixes = prefixes
+
+    def __call__(self, event_name, param, value, **kwargs):
+        """Handler that supports param values from local files."""
+        if isinstance(value, list) and len(value) == 1:
+            value = value[0]
+        try:
+            return get_paramfile(value, self._prefixes)
+        except ResourceLoadingError as e:
+            raise argprocess.ParamError(param.cli_name, str(e))
 
 
 def get_paramfile(path, cases):
