+++ tests/unit/test_clidriver.py
@@ -39,7 +39,7 @@ from awscli.clidriver import CLICommand
 from awscli.clidriver import construct_cli_error_handlers_chain
 from awscli.clidriver import ServiceCommand
 from awscli.clidriver import ServiceOperation
-from awscli.uriargumenthandler import URIArgumentHandler
+from awscli.paramfile import URIArgumentHandler
 from awscli.customizations.commands import BasicCommand
 from awscli import formatter
 from awscli.argparser import HELP_BLURB
@@ -574,7 +574,7 @@ class TestAWSCommand(BaseAWSCommandParamsTest):
         self.assertEqual(len(args_seen), 1)
         self.assertEqual(args_seen[0].unknown_arg, 'foo')
 
-    @mock.patch('awscli.uriargumenthandler.URIArgumentHandler',
+    @mock.patch('awscli.paramfile.URIArgumentHandler',
                 spec=URIArgumentHandler)
     def test_custom_arg_paramfile(self, mock_handler):
         mock_paramfile = mock.Mock(autospec=True)
@@ -604,7 +604,7 @@ class TestAWSCommand(BaseAWSCommandParamsTest):
             'file:///foo',
             mock_paramfile.call_args_list[-1][1]['value'])
 
-    @mock.patch('awscli.uriargumenthandler.URIArgumentHandler',
+    @mock.patch('awscli.paramfile.URIArgumentHandler',
                 spec=URIArgumentHandler)
     def test_custom_command_paramfile(self, mock_handler):
         mock_paramfile = mock.Mock(autospec=True)
